Mahendra_Kumar_Reddy_Kakarla/AI_contract_reviewer/requirements.txt
PyPDF2
python-docx
google-generativeai
pyyaml
asyncio
python-dotenv